Add bob-walker.py
Through an amazing oversight, we have not yet created a bob-walker in
Python. This Pull Request attempts to address this egregious error using
something that one hopes will become the reference implementation of a
bob walker.

This implements all the modes available in the Perl version, with three
notable new features:

    - If bob is being drunk and random, the amount of slurring varies. This
      makes it possible for bob to be drunk AND consistent which should
      be an allowable state.
    - bob will only say one million words before he gives up. If you'd
      like him to stop earlier, then just press CTRL-C at which point
      he's going to suggest that it might be PUB TIME!
    - If bob encounters an error, we don't care. We're just going to
      stop what we are doing and suggest having another pint. This
      probably won't stop the error occurring, but we can always fix
      that later.