Add a C version of bob-walker
Claims that the Go version of bob-walker can now be tested, by comparing
to this C implementation. To run through the 3 word list 1 million times
in C takes 0.31 seconds. With the 5 word list in sport mode, this takes
only 0.51 seconds.

Beat that Hipsters!

(Yes, I'm aware that this is probably extremely bad C, but this has
never stopped me writing code before, nor will it stop me in future)